Moved configuration.tsx into its own folder
